Move some config editor drawer code to typescript
