+++ plugins/config/src/ConfigurationEditorWidget/components/TypeSelector.tsx
@@ -2,9 +2,20 @@ import React from 'react'
 import { MenuItem, Paper, TextField } from '@mui/material'
 import { observer } from 'mobx-react'
 import { useSlotEditorStyles } from './SlotEditor'
+import { AnyConfigurationModel } from '@jbrowse/core/configuration'
 
 const TypeSelector = observer(
-  ({ typeNameChoices, slot, slotName, onChange }) => {
+  ({
+    typeNameChoices,
+    slot,
+    slotName,
+    onChange,
+  }: {
+    typeNameChoices: string[]
+    slot: AnyConfigurationModel
+    slotName: string
+    onChange: (event: React.ChangeEvent<HTMLInputElement>) => void
+  }) => {
     const { classes } = useSlotEditorStyles()
     return (
       <Paper className={classes.paper}>
