+++ products/jbrowse-web/src/tests/DrawerWidget.test.js
@@ -1,5 +1,7 @@
 import React from 'react'
 import '@testing-library/jest-dom/extend-expect'
+
+import { act } from 'react-dom/test-utils'
 import { LocalFile } from 'generic-filehandle'
 import { clearCache } from '@jbrowse/core/util/io/RemoteFileWithRangeCache'
 import { clearAdapterCache } from '@jbrowse/core/data_adapters/dataAdapterCache'
@@ -27,9 +29,11 @@ test('variant track test - opens feature detail view', async () => {
   )
   await findByText('Help')
   const view = state.session.views[0]
-  view.setNewView(0.05, 5000)
+  act(() => view.setNewView(0.05, 5000))
   fireEvent.click(await findByTestId('htsTrackEntry-volvox_filtered_vcf'))
-  view.tracks[0].displays[0].setFeatureIdUnderMouse('test-vcf-604452')
+  act(() =>
+    view.tracks[0].displays[0].setFeatureIdUnderMouse('test-vcf-604452'),
+  )
   const feats1 = await findAllByTestId('test-vcf-604452', {}, waitForOptions)
   fireEvent.click(feats1[0])
 
@@ -46,12 +50,12 @@ test('widget drawer navigation', async () => {
   const state = pm.rootModel
   const { findByTestId, findByText } = render(<JBrowse pluginManager={pm} />)
   await findByText('Help')
-  state.session.views[0].setNewView(0.05, 5000)
+  act(() => state.session.views[0].setNewView(0.05, 5000))
   // opens a config editor widget
   fireEvent.click(await findByTestId('htsTrackEntry-volvox_filtered_vcf'))
   fireEvent.click(await findByTestId('htsTrackEntryMenu-volvox_filtered_vcf'))
   fireEvent.click(await findByText('Settings'))
-  await findByTestId('configEditor')
+  await findByTestId('configEditor', {}, waitForOptions)
   // shows up when there active widgets
   fireEvent.mouseDown(
     getByRole(await findByTestId('widget-drawer-selects'), 'button'),
